requirements.txt
ursina>=0.7.0
pytest
Pillow>=9.0.0
